NomalEnemy,NoBulletEnemyの継承化
