+++ Test/Assets/Data/EnemyData/BossData.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 829d33810ff0b2444989e6020f719a74
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
