+++ Test/Assets/Data/EnemyData/NoBulletEnemyData.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 76746e7e1c37daf4989089f2a735aa68
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
