+++ Test/Assets/Data/EnemyData/NoBulletEnemyData/NoBullet Enemy Base Data.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd556043dd15be54a8631087178b4120
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
