+++ Test/Assets/Data/EnemyData/NoBulletEnemyData/NoBullet Horizontal Enemy Data.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b32a3ef600257fe4eacf18701fe6d261
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
