+++ Test/Assets/Data/EnemyData/NoBulletEnemyData/NoBullet Vertical Enemy Data.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fb19af088a3d24e4eabb97d3038b0404
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
