+++ Test/Assets/Data/EnemyData/NomalEnemyData.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f0c2cd387f2054541b65a81ff72776f7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
