+++ Test/Assets/Data/EnemyData/NomalEnemyData/Horizontal Enemy Data.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 267a4cbeb6a4c5741991cc855befde29
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
