+++ Test/Assets/Data/EnemyData/NomalEnemyData/Nomal Enemy Base Data.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c081196c69626f4b9a7a11c45150c04
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
