+++ Test/Assets/Data/EnemyData/NomalEnemyData/Vertical Enemy Data.asset
@@ -1,0 +1,39 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c7f6c916288e9244db3f9b99f1bec802, type: 3}
+  m_Name: Vertical Enemy Data
+  m_EditorClassIdentifier: 
+  maxHP: 0
+  enemyType: 0
+  coolTime: 0
+  soundData: {fileID: 0}
+  moveType: 0
+  leftRenge: 0
+  rightRenge: 0
+  moveHorizontalTime: 0
+  upRenge: 0
+  downRenge: 0
+  moveVerticalTime: 0
+  moveWaitTime: 0
+  enemyAttackType: 0
+  bulletObj: {fileID: 0}
+  bulletRate: 0
+  beforeAttackTime: 0
+  blinkDuration: 0
+  isRotate: 0
+  rotateTime: 0
+  isJump: 0
+  jumpPower: 0
+  jumpCoolTime: 0
+  groundLayer:
+    serializedVersion: 2
+    m_Bits: 0
