+++ Test/Assets/Data/EnemyData/NomalEnemyData/Vertical Enemy Data.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 62904d9c7e578f749bd7c5458a0a80cb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
