+++ Test/Assets/Prefabs/Enemy/NoBulletEnemy/NoBulletHorizontalEnemy Variant.prefab
@@ -56,6 +56,10 @@ PrefabInstance:
       propertyPath: patrolEnemyData
       value: 
       objectReference: {fileID: 11400000, guid: bfcdde73abbbf8543aa60bab7ef0dfb1, type: 2}
+    - target: {fileID: 3004321200888940903, guid: 4bfdfdd99d73f1443883c20faca1e132, type: 3}
+      propertyPath: enemyData
+      value: 
+      objectReference: {fileID: 11400000, guid: b32a3ef600257fe4eacf18701fe6d261, type: 2}
     - target: {fileID: 4931697069532253782, guid: 4bfdfdd99d73f1443883c20faca1e132, type: 3}
       propertyPath: m_Name
       value: NoBulletHorizontalEnemy Variant
