+++ Test/Assets/Prefabs/Enemy/NoBulletEnemy/NoBulletVerticalEnemy Variant.prefab
@@ -56,6 +56,10 @@ PrefabInstance:
       propertyPath: patrolEnemyData
       value: 
       objectReference: {fileID: 11400000, guid: 6a9c8c5ce8099b14b99cf12e12dcdacd, type: 2}
+    - target: {fileID: 3004321200888940903, guid: 4bfdfdd99d73f1443883c20faca1e132, type: 3}
+      propertyPath: enemyData
+      value: 
+      objectReference: {fileID: 11400000, guid: fb19af088a3d24e4eabb97d3038b0404, type: 2}
     - target: {fileID: 4931697069532253782, guid: 4bfdfdd99d73f1443883c20faca1e132, type: 3}
       propertyPath: m_Name
       value: NoBulletVerticalEnemy Variant
