+++ Test/Assets/Prefabs/Enemy/NomalEnemy/HorizontalEnemy Variant.prefab
@@ -52,6 +52,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 8786088040759193403, guid: e065fdea945b6b343aec1b4f60339d0a, type: 3}
+      propertyPath: enemyData
+      value: 
+      objectReference: {fileID: 11400000, guid: 267a4cbeb6a4c5741991cc855befde29, type: 2}
     - target: {fileID: 8837454781917888347, guid: e065fdea945b6b343aec1b4f60339d0a, type: 3}
       propertyPath: m_Enabled
       value: 1
