+++ Test/Assets/Prefabs/Enemy/NomalEnemy/VerticalEnemy Variant.prefab
@@ -52,6 +52,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 8786088040759193403, guid: e065fdea945b6b343aec1b4f60339d0a, type: 3}
+      propertyPath: enemyData
+      value: 
+      objectReference: {fileID: 11400000, guid: 62904d9c7e578f749bd7c5458a0a80cb, type: 2}
     - target: {fileID: 8837454781917888347, guid: e065fdea945b6b343aec1b4f60339d0a, type: 3}
       propertyPath: m_Enabled
       value: 1
