+++ Test/Assets/Scripts/Enemy/BossScripts/DepthBoss.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4b13aec56f9f0b845a3205353d9b2977
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
