+++ Test/Assets/Scripts/Enemy/BossScripts/MoveBoss.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 68a422c4be367df499dcb3a38180b68a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
