+++ Test/Assets/Scripts/Enemy/BossScripts/MoveBoss/Laser.cs
