+++ Test/Assets/Scripts/Enemy/BossScripts/MoveBoss/LaserAttck.cs
