+++ Test/Assets/Scripts/Enemy/BossScripts/MoveBoss/LeftVerticalMove.cs
@@ -1,11 +1,10 @@
-using System.Collections;
 using UnityEngine;
 
-public class RightVerticalMove : MonoBehaviour,IState
+public class LeftVerticalMove : MonoBehaviour,IState
 {
-    //右の縦移動
+    //左の縦移動
     private EnemyAI enemyAI;
-    public RightVerticalMove(EnemyAI enemyAI)
+    public LeftVerticalMove(EnemyAI enemyAI)
     {
         this.enemyAI = enemyAI;
     }
@@ -25,16 +24,16 @@ public class RightVerticalMove : MonoBehaviour,IState
     //移動に関する座標
     private Vector3 centerPos;
     private Vector3 startPos;
-    private Vector3 rightCenterPos;
-    private Vector3 rightUpPos;
-    private Vector3 rightDownPos;
-    private float rightPosX;
+    private Vector3 leftCenterPos;
+    private Vector3 leftUpPos;
+    private Vector3 leftDownPos;
+    private float leftPosX;
     private float upPosY;
     private float downPosY;
 
     public void Enter()
     {
-        Debug.Log("1_Enter");
+        Debug.Log("2_Enter");
         moveCounter = 0;
         moveEnemy = enemyAI.moveObj;
         enemyAI.StopAttack();
@@ -45,32 +44,32 @@ public class RightVerticalMove : MonoBehaviour,IState
 
         centerPos = enemyAI.centerPos;
         startPos = moveEnemy.transform.position;
-        rightPosX = centerPos.x + enemyAI.enemyData.rightRenge;
+        leftPosX = centerPos.x - enemyAI.enemyData.leftRenge;
         upPosY = centerPos.y + enemyAI.enemyData.upRenge;
         downPosY = centerPos.y - enemyAI.enemyData.downRenge;
-        rightUpPos = new Vector3(rightPosX, upPosY, centerPos.z);
-        rightDownPos = new Vector3(rightPosX, downPosY, centerPos.z);
-        rightCenterPos= new Vector3(rightPosX, centerPos.y, centerPos.z);
+        leftUpPos = new Vector3(leftPosX, upPosY, centerPos.z);
+        leftDownPos = new Vector3(leftPosX, downPosY, centerPos.z);
+        leftCenterPos= new Vector3(leftPosX, centerPos.y, centerPos.z);
         
         rotateAxis = enemyAI.rotateAxis;
-        rotateAxisRotate = new Vector3(0, 0, 0);
+        rotateAxisRotate = new Vector3(0, 0, 180);
         if (rotateAxis.transform.eulerAngles == rotateAxisRotate)
         {
             finishRotating = true;
         }
-        else if (rotateAxis.transform.eulerAngles == new Vector3(0, 0, 180))
+        else if (rotateAxis.transform.eulerAngles == new Vector3(0, 0, 0))
         {
-            angleZ = -90;
+            angleZ = 90;
         }
         else //if (rotateAxis.transform.eulerAngles.z == rotateAxisRotate.z)
         {
-            angleZ = -45;
+            angleZ = 45;
         }
     }
 
     public void Execute()
     {
-        //Debug.Log("1_Execute");
+        //Debug.Log("2_Execute");
         if (isCoolTime == true)
         {
             float diff = Time.time - startTime;
@@ -85,7 +84,7 @@ public class RightVerticalMove : MonoBehaviour,IState
                 isCoolTime = false;
             }
         }
-        MoveRightCenter();
+        MoveLeftCenter();
         if (finishRotating != true)
         {
             finishRotating =
@@ -99,14 +98,15 @@ public class RightVerticalMove : MonoBehaviour,IState
 
     public void Exit()
     {
-        Debug.Log("1_Exit");
+        Debug.Log("2_Exit");
     }
 
     void Initialization()
     {
         startTime = Time.time;
+        //centerPos = moveEnemy.transform.position;
     }
-
+    
     void NextMove()
     {
         moveCounter++;
@@ -118,23 +118,7 @@ public class RightVerticalMove : MonoBehaviour,IState
         }
     }
 
-    void Rotate(Vector3 angles)
-    {
-        //Debug.Log("angle:"+rotateAxis.transform.rotation.eulerAngles);
-        if (finishRotating == true)
-        {
-            return;
-        }
-        t += Time.deltaTime;
-        rotateAxis.transform.Rotate(0, 0, angleZ * Time.deltaTime);
-        if (t >= 2.0f)
-        {
-            rotateAxis.transform.eulerAngles = angles;
-            finishRotating = true;
-        }
-    }
-    
-    void MoveRightCenter()
+    void MoveLeftCenter()
     {
         //Debug.Log("MoveRightCenter");
         if (finishMoving == true) {return;}
@@ -142,28 +126,28 @@ public class RightVerticalMove : MonoBehaviour,IState
         switch (moveCounter)
         {
             case 0:
-                if (enemyAI.EnemyMove(moveEnemy, startPos, rightCenterPos, enemyAI.enemyData.moveVerticalTime,
+                if (enemyAI.EnemyMove(moveEnemy, startPos, leftCenterPos, enemyAI.enemyData.moveVerticalTime,
                         startTime) == true)
                 {
                     NextMove();
                 }
                 break;
             case 1:
-                if (enemyAI.EnemyMove(moveEnemy, rightCenterPos, rightUpPos, enemyAI.enemyData.moveHorizontalTime,
+                if (enemyAI.EnemyMove(moveEnemy, leftCenterPos, leftUpPos, enemyAI.enemyData.moveHorizontalTime,
                         startTime) == true)
                 {
                     NextMove();
                 }
                 break;
             case 2:
-                if (enemyAI.EnemyMove(moveEnemy, rightUpPos, rightDownPos, enemyAI.enemyData.moveHorizontalTime * 2,
+                if (enemyAI.EnemyMove(moveEnemy, leftUpPos, leftDownPos, enemyAI.enemyData.moveHorizontalTime * 2,
                         startTime) == true)
                 {
                     NextMove();
                 }
                 break;
             case 3:
-                if (enemyAI.EnemyMove(moveEnemy, rightDownPos, rightCenterPos, enemyAI.enemyData.moveHorizontalTime,
+                if (enemyAI.EnemyMove(moveEnemy, leftDownPos, leftCenterPos, enemyAI.enemyData.moveHorizontalTime,
                         startTime) == true)
                 {
                     NextMove();
