+++ Test/Assets/Scripts/Enemy/NomalEnemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ebe580c44485c084fb6f99d781bdac80
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
