+++ Test/Assets/Scripts/Enemy/NomalEnemy/NomalEnemy.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+using UnityEditor;
+
+#if UNITY_EDITOR
+[CustomEditor(typeof(EnemyAI))]
+#endif
+
+public class NomalEnemy : EnemyAI
+{
+    override protected void Update()
+    {
+        base.Update();
+        //stateMachine.Update();
+    }
+}
